Added ServerMetricsTestUtils.
